+++ src/components/Dropdown/DropdownContent/index.ts
@@ -1,0 +1,2 @@
+export { DropdownContent } from './DropdownContent';
+export { DropdownMenuContentProps as DropdownContentProps } from '@radix-ui/react-dropdown-menu';
