+++ src/components/Dropdown/DropdownSubContent/index.ts
@@ -1,0 +1,2 @@
+export { DropdownSubContent } from './DropdownSubContent';
+export { DropdownMenuSubContentProps as DropdownSubContentProps } from '@radix-ui/react-dropdown-menu';
