+++ src/components/Dropdown/DropdownSubTrigger/index.ts
@@ -1,2 +1,2 @@
 export { DropdownSubTrigger } from './DropdownSubTrigger';
-export { DropdownMenuSubContentProps as DropdownSubTriggerProps } from '@radix-ui/react-dropdown-menu';
+export { DropdownMenuSubTriggerProps as DropdownSubTriggerProps } from '@radix-ui/react-dropdown-menu';
