--- src/components/Dropdown/DropdownTrigger.tsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import { Trigger, DropdownMenuTriggerProps } from '@radix-ui/react-dropdown-menu';
-
-export const DropdownTrigger = (props: DropdownMenuTriggerProps) => <Trigger asChild {...props} />;
