+++ src/components/Dropdown/DropdownTrigger/index.ts
@@ -1,0 +1,2 @@
+export { DropdownTrigger } from './DropdownTrigger';
+export { DropdownMenuTriggerProps as DropdownTriggerProps } from '@radix-ui/react-dropdown-menu';
